refactor out handleMarker
